+++ uni_modules/uview-ui/components/u-code-input/u-code-input.vue
@@ -23,6 +23,9 @@
 				v-if="mode === 'line'"
 				:style="[lineStyle]"
 			></view>
+			<!-- #ifndef APP-PLUS -->
+			<view v-if="isFocus && codeArray.length === index" :style="{backgroundColor: color}" class="u-code-input__item__cursor"></view>
+			<!-- #endif -->
 		</view>
 		<input
 			:disabled="disabledKeyboard"
@@ -30,11 +33,14 @@
 			:focus="focus"
 			:value="inputValue"
 			:maxlength="maxlength"
+			:adjustPosition="adjustPosition"
 			class="u-code-input__input"
 			@input="inputHandler"
 			:style="{
 				height: $u.addUnit(size) 
 			}"
+			@focus="isFocus = true"
+			@blur="isFocus = false"
 		/>
 	</view>
 </template>
@@ -69,7 +75,8 @@
 		mixins: [uni.$u.mpMixin, uni.$u.mixin, props],
 		data() {
 			return {
-				inputValue: ''
+				inputValue: '',
+				isFocus: this.focus
 			}
 		},
 		watch: {
@@ -166,6 +173,10 @@
 
 <style lang="scss" scoped>
 	@import "../../libs/css/components.scss";
+	$u-code-input-cursor-width: 1px;
+	$u-code-input-cursor-height: 40%;
+	$u-code-input-cursor-animation-duration: 1s;
+	$u-code-input-cursor-animation-name: u-cursor-flicker;
 
 	.u-code-input {
 		@include flex;
@@ -176,6 +187,7 @@
 			@include flex;
 			justify-content: center;
 			align-items: center;
+			position: relative;
 
 			&__text {
 				font-size: 15px;
@@ -197,6 +209,18 @@
 				width: 40px;
 				background-color: $u-content-color;
 			}
+			/* #ifndef APP-PLUS */
+			&__cursor {
+				position: absolute;
+				top: 50%;
+				left: 50%;
+				transform: translate(-50%,-50%);
+				width: $u-code-input-cursor-width;
+				height: $u-code-input-cursor-height;
+				animation: $u-code-input-cursor-animation-duration u-cursor-flicker infinite;
+			}
+			/* #endif */
+			
 		}
 
 		&__input {
@@ -210,4 +234,19 @@
 			text-align: left;
 		}
 	}
+	
+	/* #ifndef APP-PLUS */
+	@keyframes u-cursor-flicker {
+		0% {
+		    opacity: 0;
+		}
+		50% {
+		    opacity: 1;
+		}
+		100% {
+		    opacity: 0;
+		}
+	}
+	/* #endif */
+
 </style>