+++ uni_modules/uview-ui/libs/config/props/tabs.js
@@ -21,6 +21,7 @@ export default {
         }),
         lineWidth: 20,
         lineHeight: 3,
+        lineBgSize: 'cover',
         itemStyle: () => ({
             height: '44px'
         }),